requirements.txt
pytest
pytest-playwright
playwright
pytest-html 
pytest-xdist  
pytest-html 
pytest-xdist 
requests
jsonschema
python-dotenv
faker
allure-pytest 